Fix Helper user info auth (#2179)
Helper no longer has separate widget and user info HMAC secrets, but we
still have an old key in a separate config variable which is causing
user info requests to fail (as reported by support).

This updates the endpoint to use the correct secret and removes the
config for the old one.